move chart selectors
