+++ x-pack/plugins/apm/public/components/shared/charts/metrics_chart/index.tsx
